+++ packages/core/tracing/src/utils/lifecycle.ts
@@ -1,8 +1,8 @@
-import { MarkerType, type Edge } from '@vue-flow/core'
-import { KONG_PHASES, LifecycleNodeType, SPAN_NAMES } from '../constants'
+import { KUI_COLOR_BACKGROUND_DISABLED } from '@kong/design-tokens'
+import { MarkerType } from '@vue-flow/core'
+import { KONG_PHASES, LifecycleNodeType, NODE_STRIPE_COLOR_MAPPING, NODE_STRIPE_COLOR_MAPPING_START_EXP, SPAN_NAMES } from '../constants'
 import type { LifecycleGraph, LifecycleNodeData, SpanNode } from '../types'
 import { getPhaseAndPlugin } from './spans'
-import { getDurationFormatter } from './time'
 
 /**
  * Traverse a plugin span tree and build the {@link LifecycleNodeData} for the plugin and accumulate
@@ -14,7 +14,7 @@ import { getDurationFormatter } from './time'
  * @param node The root node of the plugin span tree
  * @returns the node data and the duration that has been excluded
  */
-const buildPluginNodeData = (node: SpanNode): [LifecycleNodeData, number] | undefined => {
+const buildPluginNodeData = (node: SpanNode): [LifecycleNodeData & { id: string }, number] | undefined => {
   // Try to parse the phase and plugin name from the span name
   const pluginSpan = getPhaseAndPlugin(node.span.name)
   if (!pluginSpan || pluginSpan.suffix) {
@@ -61,6 +61,7 @@ const buildPluginNodeData = (node: SpanNode): [LifecycleNodeData, number] | unde
   }
 
   return [{
+    id: `${pluginSpan.plugin}:${pluginSpan.phase}`, // For reference and node generation only
     label: pluginSpan.plugin,
     type: nodeType,
     badge: pluginSpan.phase,
@@ -70,11 +71,10 @@ const buildPluginNodeData = (node: SpanNode): [LifecycleNodeData, number] | unde
 }
 
 export const buildLifecycleGraph = (root: SpanNode): LifecycleGraph => {
-  const fmt = getDurationFormatter()
   const ingressSpans: SpanNode[] = []
   const egressSpans: SpanNode[] = []
-  const requestNodesData: LifecycleNodeData[] = []
-  const responseNodesData: LifecycleNodeData[] = []
+  const requestNodesData: (LifecycleNodeData & { id: string })[] = []
+  const responseNodesData: (LifecycleNodeData & { id: string })[] = []
   const upstreamSpans: SpanNode[] = []
 
   let pluginDurationExcluded = 0
@@ -122,62 +122,233 @@ export const buildLifecycleGraph = (root: SpanNode): LifecycleGraph => {
     }
   }
 
-  const graph: LifecycleGraph = {
-    nodes: [],
+  const graph: Omit<LifecycleGraph, 'nodes'> = {
+    nodeTree: {
+      client: {
+        node: {
+          id: LifecycleNodeType.CLIENT,
+          position: { x: 0, y: 0 },
+          data: {
+            label: 'Client',
+            type: LifecycleNodeType.CLIENT,
+            showSourceHandle: true,
+          },
+          zIndex: 2,
+        },
+        in: {
+          id: LifecycleNodeType.CLIENT_IN,
+          position: { x: 0, y: 0 },
+          data: {
+            type: LifecycleNodeType.CLIENT_IN,
+            durationNano: ingressSpans.reduce((duration, span) => duration + (span.durationNano ?? 0), 0) + pluginDurationExcluded,
+            durationTooltipKey: 'lifecycle.client_in.tooltip',
+          },
+          zIndex: 10, // Because we may want to show tooltips on this node
+        },
+        out: {
+          id: LifecycleNodeType.CLIENT_OUT,
+          position: { x: 0, y: 0 },
+          data: {
+            type: LifecycleNodeType.CLIENT_OUT,
+            durationNano: egressSpans.reduce((duration, span) => duration + (span.durationNano ?? 0), 0),
+            durationTooltipKey: 'lifecycle.client_out.tooltip',
+          },
+          zIndex: 10, // Because we may want to show tooltips on this node
+        },
+      },
+      requests: requestNodesData.length > 0 ? {
+        node: {
+          id: LifecycleNodeType.REQUEST_GROUP,
+          position: { x: 0, y: 0 },
+          data: {
+            label: 'Requests',
+            type: LifecycleNodeType.REQUEST_GROUP,
+            labelPlacement: 'top left',
+          },
+          zIndex: 1,
+        },
+        children: requestNodesData.map((nodeData, i) => ({
+          id: `${LifecycleNodeType.REQUEST}:${nodeData.id}`,
+          position: { x: 0, y: 0 },
+          parentNode: LifecycleNodeType.REQUEST_GROUP,
+          data: {
+            ...nodeData,
+            showTargetHandle: i > 0,
+            showSourceHandle: true,
+          },
+          zIndex: 1,
+        })),
+      } : undefined,
+      upstream: {
+        id: LifecycleNodeType.UPSTREAM,
+        position: { x: 0, y: 0 },
+        data: {
+          label: 'Upstream',
+          type: LifecycleNodeType.UPSTREAM,
+          durationNano: upstreamSpans.reduce((duration, span) => duration + (span.durationNano ?? 0), 0),
+          showSourceHandle: true,
+        },
+        zIndex: 1,
+      },
+      responses: responseNodesData.length > 0 ? {
+        node: {
+          id: LifecycleNodeType.RESPONSE_GROUP,
+          position: { x: 0, y: 0 },
+          data: {
+            label: 'Responses',
+            type: LifecycleNodeType.RESPONSE_GROUP,
+            labelPlacement: 'bottom left',
+          },
+          zIndex: 1,
+        },
+        children: responseNodesData.map((nodeData, i) => ({
+          id: `${LifecycleNodeType.RESPONSE}:${nodeData.id}`,
+          position: { x: 0, y: 0 },
+          parentNode: LifecycleNodeType.RESPONSE_GROUP,
+          data: {
+            ...nodeData,
+            showTargetHandle: i > 0,
+            showSourceHandle: true,
+          },
+          zIndex: 1,
+        })),
+      } : undefined,
+    },
     edges: [],
   }
 
-  graph.nodes.push({
-    id: 'client',
-    position: { x: 0, y: 0 },
-    data: {
-      label: 'Client',
-      type: LifecycleNodeType.CLIENT,
-    },
+  const clientNodes = graph.nodeTree.client
+
+  graph.edges.push({
+    id: `${clientNodes.node.id}->${clientNodes.out.id}`,
+    source: clientNodes.node.id,
+    target: clientNodes.out.id,
+    type: 'seamless-smoothstep',
+    zIndex: 1,
   })
 
-  requestNodesData.forEach((nodeData, i) => {
-    graph.nodes.push({
-      id: `request#${i}`,
-      position: { x: 0, y: 0 },
-      data: nodeData,
-    })
-  })
+  if (graph.nodeTree.requests && graph.nodeTree.requests.children.length > 0) {
+    const children = graph.nodeTree.requests.children
 
-  graph.nodes.push({
-    id: 'upstream',
-    position: { x: 0, y: 0 },
-    data: {
-      label: 'Upstream',
-      type: LifecycleNodeType.UPSTREAM,
-      durationNano: upstreamSpans.reduce((duration, span) => duration + (span.durationNano ?? 0), 0),
-    },
-  })
-
-  responseNodesData.forEach((nodeData, i) => {
-    graph.nodes.push({
-      id: `response#${i}`,
-      position: { x: 0, y: 0 },
-      data: nodeData,
-    })
-  })
-
-  for (let i = 0; i < graph.nodes.length; i++) {
-    const j = i < graph.nodes.length - 1 ? i + 1 : 0
-    const edge: Edge = {
-      id: `${graph.nodes[i].id}->${graph.nodes[j].id}`,
-      source: graph.nodes[i].id,
-      target: graph.nodes[j].id,
-      type: 'smoothstep',
+    graph.edges.push({
+      id: `${clientNodes.out.id}->${children[0].id}`,
+      source: clientNodes.out.id,
+      target: children[0].id,
+      type: 'seamless-smoothstep',
       markerEnd: MarkerType.ArrowClosed,
+      zIndex: 2,
+    })
+
+    for (let i = 0; i < children.length - 1; i++) {
+      graph.edges.push({
+        id: `${children[i].id}->${children[i + 1].id}`,
+        source: children[i].id,
+        target: children[i + 1].id,
+        type: 'smoothstep', // Using built-in edge type because handles are available on both side–we don't care about the gaps
+        zIndex: 2,
+      })
     }
-    if (i === 0) {
-      edge.label = fmt(ingressSpans.reduce((duration, span) => duration + (span.durationNano ?? 0), 0) + pluginDurationExcluded)
-    } else if (i === graph.nodes.length - 1) {
-      edge.label = fmt(egressSpans.reduce((duration, span) => duration + (span.durationNano ?? 0), 0))
-    }
-    graph.edges.push(edge)
+
+    graph.edges.push({
+      id: `${children[children.length - 1].id}->${graph.nodeTree.upstream.id}`,
+      source: children[children.length - 1].id,
+      target: graph.nodeTree.upstream.id,
+      type: 'seamless-smoothstep',
+      markerEnd: MarkerType.ArrowClosed,
+      zIndex: 2,
+    })
+  } else {
+    graph.edges.push(
+      {
+        id: `${clientNodes.out.id}->${graph.nodeTree.upstream.id}`,
+        source: clientNodes.out.id,
+        target: graph.nodeTree.upstream.id,
+        type: 'seamless-smoothstep',
+        markerEnd: MarkerType.ArrowClosed,
+        zIndex: 1,
+      },
+    )
   }
 
-  return graph
+  if (graph.nodeTree.responses) {
+    const children = graph.nodeTree.responses.children
+
+    graph.edges.push({
+      id: `${graph.nodeTree.upstream.id}->${children[0].id}`,
+      source: graph.nodeTree.upstream.id,
+      target: children[0].id,
+      type: 'seamless-smoothstep',
+      markerEnd: MarkerType.ArrowClosed,
+      zIndex: 2,
+    })
+
+    for (let i = 0; i < children.length - 1; i++) {
+      graph.edges.push({
+        id: `${children[i].id}->${children[i + 1].id}`,
+        source: children[i].id,
+        target: children[i + 1].id,
+        type: 'smoothstep', // Using built-in edge type because handles are available on both side–we don't care about the gaps
+        zIndex: 2,
+      })
+    }
+
+    graph.edges.push({
+      id: `${children[children.length - 1].id}->${clientNodes.in.id}`,
+      source: children[children.length - 1].id,
+      target: clientNodes.in.id,
+      type: 'seamless-smoothstep',
+      zIndex: 2,
+    })
+  } else {
+    graph.edges.push({
+      id: `${graph.nodeTree.upstream.id}->${clientNodes.in.id}`,
+      source: graph.nodeTree.upstream.id,
+      target: clientNodes.in.id,
+      type: 'seamless-smoothstep',
+      zIndex: 1,
+    })
+  }
+
+  graph.edges.push({
+    id: `${clientNodes.in.id}->${clientNodes.node.id}`,
+    source: clientNodes.in.id,
+    target: clientNodes.node.id,
+    type: 'seamless-smoothstep',
+    markerEnd: MarkerType.ArrowClosed,
+    zIndex: 1,
+  })
+
+  return {
+    ...graph,
+    nodes: [
+      clientNodes.node,
+      clientNodes.out,
+      ...graph.nodeTree.requests ? [
+        graph.nodeTree.requests.node,
+        ...graph.nodeTree.requests.children,
+      ] : [],
+      graph.nodeTree.upstream,
+      ...graph.nodeTree.responses ? [
+        graph.nodeTree.responses.node,
+        ...graph.nodeTree.responses.children,
+      ] : [],
+      clientNodes.in,
+    ],
+  }
+}
+
+/**
+ * See {@link NODE_STRIPE_COLOR_MAPPING}
+ */
+export const getNodeStripeColor = (durationNano?: number) => {
+  if (durationNano === undefined || durationNano < 0) {
+    return KUI_COLOR_BACKGROUND_DISABLED
+  }
+
+  return NODE_STRIPE_COLOR_MAPPING[
+    Math.min(
+      NODE_STRIPE_COLOR_MAPPING.length - 1,
+      Math.max(0, Math.floor(Math.log10(durationNano)) - NODE_STRIPE_COLOR_MAPPING_START_EXP),
+    )
+  ]
 }