+++ packages/entities/entities-consumers/sandbox/pages/ConsumerListPage.vue
@@ -2,6 +2,9 @@
   <SandboxPermissionsControl
     @update="handlePermissionsUpdate"
   />
+  <h2>Konnect Actions Outside</h2>
+  <div id="kong-ui-app-page-header-action-button" />
+
   <h2>Konnect API</h2>
   <ConsumerList
     v-if="permissions"
@@ -12,6 +15,7 @@
     :can-edit="permissions.canEdit"
     :can-retrieve="permissions.canRetrieve"
     :config="konnectConfig"
+    use-action-outside
     @add:success="onAddSuccess"
     @copy:error="onCopyIdError"
     @copy:success="onCopyIdSuccess"