+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.6.22](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.6.21...@kong-ui-public/analytics-metric-provider@8.6.22) (2025-02-17)
+
+
+### Bug Fixes
+
+* **deps:** update all [@kong](https://github.com/kong) scoped dependencies (minor) ([#1961](https://github.com/Kong/public-ui-components/issues/1961)) ([7da943a](https://github.com/Kong/public-ui-components/commit/7da943a33a3c2325cbcd6bf213828ae9f4c22b60))
+
+
+
+
+
+## [8.6.21](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.6.20...@kong-ui-public/analytics-metric-provider@8.6.21) (2025-02-13)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [8.6.20](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.6.19...@kong-ui-public/analytics-metric-provider@8.6.20) (2025-02-11)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [8.6.19](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.6.18...@kong-ui-public/analytics-metric-provider@8.6.19) (2025-01-27)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-metric-provider