+++ packages/core/forms/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.10.8](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/forms@4.10.7...@kong-ui-public/forms@4.10.8) (2025-02-17)
+
+
+### Bug Fixes
+
+* **deps:** update all [@kong](https://github.com/kong) scoped dependencies (minor) ([#1961](https://github.com/Kong/public-ui-components/issues/1961)) ([7da943a](https://github.com/Kong/public-ui-components/commit/7da943a33a3c2325cbcd6bf213828ae9f4c22b60))
+
+
+
+
+
 ## [4.10.7](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/forms@4.10.6...@kong-ui-public/forms@4.10.7) (2025-01-28)