+++ packages/entities/entities-consumer-groups/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.6.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumer-groups@3.6.0...@kong-ui-public/entities-consumer-groups@3.6.1) (2025-02-17)
+
+
+### Bug Fixes
+
+* **deps:** update all [@kong](https://github.com/kong) scoped dependencies (minor) ([#1961](https://github.com/Kong/public-ui-components/issues/1961)) ([7da943a](https://github.com/Kong/public-ui-components/commit/7da943a33a3c2325cbcd6bf213828ae9f4c22b60))
+
+
+
+
+
+# [3.6.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumer-groups@3.5.74...@kong-ui-public/entities-consumer-groups@3.6.0) (2025-02-14)
+
+
+### Features
+
+* **consumer-groups:** new empty state [khcp-14827] ([#1936](https://github.com/Kong/public-ui-components/issues/1936)) ([1443167](https://github.com/Kong/public-ui-components/commit/1443167185e9a2bf9c4acca02220e0d8a7981a73))
+
+
+
+
+
 ## [3.5.74](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumer-groups@3.5.73...@kong-ui-public/entities-consumer-groups@3.5.74) (2025-02-03)
 
 **Note:** Version bump only for package @kong-ui-public/entities-consumer-groups