+++ packages/entities/entities-plugins/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.10.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.10.4...@kong-ui-public/entities-plugins@9.10.5) (2025-02-17)
+
+
+### Bug Fixes
+
+* **deps:** update all [@kong](https://github.com/kong) scoped dependencies (minor) ([#1961](https://github.com/Kong/public-ui-components/issues/1961)) ([7da943a](https://github.com/Kong/public-ui-components/commit/7da943a33a3c2325cbcd6bf213828ae9f4c22b60))
+
+
+
+
+
+## [9.10.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.10.3...@kong-ui-public/entities-plugins@9.10.4) (2025-02-14)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+## [9.10.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.10.2...@kong-ui-public/entities-plugins@9.10.3) (2025-02-14)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+## [9.10.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.10.1...@kong-ui-public/entities-plugins@9.10.2) (2025-02-11)
+
+
+### Bug Fixes
+
+* ff guard v2 empty state changes ([#1947](https://github.com/Kong/public-ui-components/issues/1947)) ([b408aea](https://github.com/Kong/public-ui-components/commit/b408aeaf7124673ab288409eee9151871e3b8c87))
+
+
+
+
+
+## [9.10.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.10.0...@kong-ui-public/entities-plugins@9.10.1) (2025-02-11)
+
+
+### Bug Fixes
+
+* **entities-plugins:** fields set to null ([#1939](https://github.com/Kong/public-ui-components/issues/1939)) ([a400abf](https://github.com/Kong/public-ui-components/commit/a400abf8a5fa39d6b0288952784373cc9cb2a22d))
+
+
+
+
+
 # [9.10.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.9.9...@kong-ui-public/entities-plugins@9.10.0) (2025-02-07)